requirements.txt
# ==========================
# CORE DJANGO
# ==========================
Django>=4.2
djangorestframework>=3.14
python-dotenv>=1.0

# ==========================
# DATABASE (MONGODB)
# ==========================
pymongo>=4.6
dnspython>=2.4

# ==========================
# DATA & ML
# ==========================
pandas>=2.1
numpy>=1.26
scikit-learn>=1.3
imbalanced-learn>=0.11
joblib>=1.3

# ==========================
# MODEL SERIALIZATION
# ==========================
pickle-mixin>=1.0

# ==========================
# PIPELINE ORCHESTRATION
# ==========================
antigravity

# ==========================
# DEPLOYMENT
# ==========================
gunicorn>=21.2
whitenoise>=6.6
